[MLG-1] Removed JSON mapping. Moved to core package.
