+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/model/Clouds.java
@@ -1,0 +1,15 @@
+
+package com.training.weather.ingestor.core.model;
+
+public class Clouds {
+
+  private Integer volume;
+
+  public Integer getVolume() {
+    return volume;
+  }
+
+  public void setVolume(Integer volume) {
+    this.volume = volume;
+  }
+}
