+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/model/Forecast.java
@@ -1,0 +1,78 @@
+package com.training.weather.ingestor.core.model;
+
+import java.util.List;
+
+public class Forecast {
+  private Integer timestamp;
+  private MainParameters mainParameters;
+  private List<Weather> weather;
+  private Clouds clouds;
+  private Wind wind;
+  private Rain rain;
+  private Snow snow;
+  private String date;
+
+  public Integer getTimestamp() {
+    return timestamp;
+  }
+
+  public void setTimestamp(Integer timestamp) {
+    this.timestamp = timestamp;
+  }
+
+  public MainParameters getMainParameters() {
+    return mainParameters;
+  }
+
+  public void setMainParameters(MainParameters mainParameters) {
+    this.mainParameters = mainParameters;
+  }
+
+  public List<Weather> getWeather() {
+    return weather;
+  }
+
+  public void setWeather(List<Weather> weather) {
+    this.weather = weather;
+  }
+
+  public Clouds getClouds() {
+    return clouds;
+  }
+
+  public void setClouds(Clouds clouds) {
+    this.clouds = clouds;
+  }
+
+  public Wind getWind() {
+    return wind;
+  }
+
+  public void setWind(Wind wind) {
+    this.wind = wind;
+  }
+
+  public Rain getRain() {
+    return rain;
+  }
+
+  public void setRain(Rain rain) {
+    this.rain = rain;
+  }
+
+  public Snow getSnow() {
+    return snow;
+  }
+
+  public void setSnow(Snow snow) {
+    this.snow = snow;
+  }
+
+  public String getDate() {
+    return date;
+  }
+
+  public void setDate(String date) {
+    this.date = date;
+  }
+}
