+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/model/MainParameters.java
@@ -1,38 +1,14 @@
+package com.training.weather.ingestor.core.model;
 
-package com.training.weather.ingestor.infrastructure.entity;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "temp",
-        "temp_min",
-        "temp_max",
-        "pressure",
-        "sea_level",
-        "grnd_level",
-        "humidity",
-        "temp_kf"})
 public class MainParameters {
 
-  @JsonProperty("temp")
   private Double temperature;
-  @JsonProperty("temp_min")
   private Double minTemperature;
-  @JsonProperty("temp_max")
   private Double maxTemperature;
-  @JsonProperty("pressure")
   private Double pressure;
-  @JsonProperty("sea_level")
   private Double seaLevelPressure;
-  @JsonProperty("grnd_level")
   private Double groundLevelPressure;
-  @JsonProperty("humidity")
   private Integer humidity;
-  @JsonProperty("temp_kf")
-  private Integer tempKf;
 
   /**
    * Creates instance of {@link MainParameters}.
@@ -67,7 +43,6 @@ public class MainParameters {
     this.seaLevelPressure = seaLevelPressure;
     this.groundLevelPressure = groundLevelPressure;
     this.humidity = humidity;
-    this.tempKf = tempKf;
   }
 
   public Double getTemperature() {
@@ -124,13 +99,5 @@ public class MainParameters {
 
   public void setHumidity(Integer humidity) {
     this.humidity = humidity;
-  }
-
-  public Integer getTempKf() {
-    return tempKf;
-  }
-
-  public void setTempKf(Integer tempKf) {
-    this.tempKf = tempKf;
   }
 }
