+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/model/Rain.java
@@ -1,0 +1,15 @@
+
+package com.training.weather.ingestor.core.model;
+
+public class Rain {
+
+  private double volume;
+
+  public double getVolume() {
+    return volume;
+  }
+
+  public void setVolume(double volume) {
+    this.volume = volume;
+  }
+}
