+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/model/Snow.java
@@ -1,0 +1,14 @@
+package com.training.weather.ingestor.core.model;
+
+public class Snow {
+
+  private String volume;
+
+  public String getVolume() {
+    return volume;
+  }
+
+  public void setVolume(String volume) {
+    this.volume = volume;
+  }
+}
