+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/model/Weather.java
@@ -1,25 +1,11 @@
 
-package com.training.weather.ingestor.infrastructure.entity;
+package com.training.weather.ingestor.core.model;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "id",
-        "main",
-        "description",
-        "icon"})
 public class Weather {
 
-  @JsonProperty("id")
   private Integer id;
-  @JsonProperty("main")
   private String main;
-  @JsonProperty("description")
   private String description;
-  @JsonProperty("icon")
   private String icon;
 
   /**
