+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/model/Wind.java
@@ -1,0 +1,36 @@
+
+package com.training.weather.ingestor.core.model;
+
+public class Wind {
+
+  private Double speed;
+  private Double degree;
+
+  /**
+   * Creates instance of {@link Wind}.
+   */
+  public Wind() {
+    //Default Constructor.
+  }
+
+  public Wind(Double speed, Double degree) {
+    this.speed = speed;
+    this.degree = degree;
+  }
+
+  public Double getSpeed() {
+    return speed;
+  }
+
+  public void setSpeed(Double speed) {
+    this.speed = speed;
+  }
+
+  public Double getDegree() {
+    return degree;
+  }
+
+  public void setDegree(Double degree) {
+    this.degree = degree;
+  }
+}
