--- weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/infrastructure/entity/Rain.java
@@ -1,33 +1,0 @@
-
-package com.training.weather.ingestor.infrastructure.entity;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "3h"})
-public class Rain {
-  @JsonProperty("3h")
-  private double volume;
-
-  /**
-   * Creates instance of {@link Rain}.
-   */
-  public Rain() {
-    //Default Constructor.
-  }
-
-  public Rain(double volume) {
-    this.volume = volume;
-  }
-
-  public double getVolume() {
-    return volume;
-  }
-
-  public void setVolume(double volume) {
-    this.volume = volume;
-  }
-}
