--- weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/infrastructure/entity/Snow.java
@@ -1,32 +1,0 @@
-package com.training.weather.ingestor.infrastructure.entity;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "3h"})
-public class Snow {
-  @JsonProperty("3h")
-  private String volume;
-
-  /**
-   * Creates instance of {@link Snow}.
-   */
-  public Snow() {
-    //Default Constructor.
-  }
-
-  public Snow(String volume) {
-    this.volume = volume;
-  }
-
-  public String getVolume() {
-    return volume;
-  }
-
-  public void setVolume(String volume) {
-    this.volume = volume;
-  }
-}
